refactor: divide extension function file
